Add ability to Sync single Vm entity
Unified interface to update single Vm via SyncAdapter#syncVm.
This method fetches fresh data from REST api + executes all configured triggers
and persists new data to database. This also required the Vm entity to be
extended to hold additional display data that were only transient before.

This API should be used as single entry point for refreshing VMs and all UI
components should be bound to ContentProvider via the existing infra of
CursorLoaders. Single entry point has the advantage of executing all business
logic as a part of "entity sync" and also that all updated data is persisted -
which was not the case for manual invocations of OVirtClient from the UI layer.

Unified "sync" entry point is also a necessary precondition for future addition
of push where we want to trigger sync only for the entity that has changed (and
run all trigger + persist logic).

Future use of OVirtClient in the UI layer *for the purpose of fetching remote
state* should be discouraged and the UI-driven refresh of remote state should
follow the following architecture:

                          +---------------------+
           Triggers Sync  |                     |  Persists updates
         +--------------> |     SyncAdapter     +---------------+
         |                |                     |               |
         |                +---------------------+               |
         |                                                      |
         |                                                      |
         |                                                      v
+--------+---------------+                         +-----------------------+
|                        |   Causes cursor reload  |                       |
|  UI Activity/Fragment  | <-----------------------+    ContentProvider    |
|                        |                         |                       |
+------------------------+                         +-----------------------+